+++ tests/e2e/cases/generated_c_dce_const/in.ts
@@ -204,6 +204,9 @@ const unused_promise_resolve_object_values_object_element_call = Promise.resolve
 const unused_promise_resolve_object_values_array_join_call = Promise.resolve(Object.values(["dead_promise_resolve_object_values_array_join"]).join("|"));
 const unused_promise_resolve_object_values_string_to_string_call = Promise.resolve(Object.values("dead_promise_resolve_object_values_string_to_string").toString());
 const unused_promise_resolve_object_values_object_join_call = Promise.resolve(Object.values({ dead_promise_resolve_object_values_object_join: "dead_promise_resolve_object_values_object_join" }).join("|"));
+const unused_promise_resolve_object_values_assign_join_call = Promise.resolve(Object.values(Object.assign({} as { dead_promise_resolve_object_values_assign_join: string }, { dead_promise_resolve_object_values_assign_join: "dead_promise_resolve_object_values_assign_join" })).join("|"));
+const unused_promise_resolve_object_values_from_entries_join_call = Promise.resolve(Object.values(Object.fromEntries<{ dead_promise_resolve_object_values_from_entries_join: string }>([["dead_promise_resolve_object_values_from_entries_join", "dead_promise_resolve_object_values_from_entries_join"]])).join("|"));
+const unused_promise_resolve_object_values_entries_from_entries_join_call = Promise.resolve(Object.values(Object.fromEntries<{ dead_promise_resolve_object_values_entries_from_entries_join: string }>(Object.entries({ dead_promise_resolve_object_values_entries_from_entries_join: "dead_promise_resolve_object_values_entries_from_entries_join" }))).join("|"));
 const unused_promise_resolve_object_entries_key_call = Promise.resolve(Object.entries({ dead_promise_resolve_object_entries_key: "dead_promise_resolve_object_entries_value" })[0][0]);
 const unused_promise_resolve_object_entries_value_call = Promise.resolve(Object.entries({ dead_promise_resolve_object_entries_value_key: "dead_promise_resolve_object_entries_value" })[0][1]);
 const unused_promise_resolve_object_entries_array_value_call = Promise.resolve(Object.entries(["dead_promise_resolve_object_entries_array_value"])[0][1]);
@@ -557,6 +560,9 @@ Promise.resolve(Object.values({ top_level_dead_promise_resolve_object_values_obj
 Promise.resolve(Object.values(["top_level_dead_promise_resolve_object_values_array_join"]).join("|"));
 Promise.resolve(Object.values("top_level_dead_promise_resolve_object_values_string_to_string").toString());
 Promise.resolve(Object.values({ top_level_dead_promise_resolve_object_values_object_join: "top_level_dead_promise_resolve_object_values_object_join" }).join("|"));
+Promise.resolve(Object.values(Object.assign({} as { top_level_dead_promise_resolve_object_values_assign_join: string }, { top_level_dead_promise_resolve_object_values_assign_join: "top_level_dead_promise_resolve_object_values_assign_join" })).join("|"));
+Promise.resolve(Object.values(Object.fromEntries<{ top_level_dead_promise_resolve_object_values_from_entries_join: string }>([["top_level_dead_promise_resolve_object_values_from_entries_join", "top_level_dead_promise_resolve_object_values_from_entries_join"]])).join("|"));
+Promise.resolve(Object.values(Object.fromEntries<{ top_level_dead_promise_resolve_object_values_entries_from_entries_join: string }>(Object.entries({ top_level_dead_promise_resolve_object_values_entries_from_entries_join: "top_level_dead_promise_resolve_object_values_entries_from_entries_join" }))).join("|"));
 Promise.resolve(Object.entries({ top_level_dead_promise_resolve_object_entries_key: "top_level_dead_promise_resolve_object_entries_value" })[0][0]);
 Promise.resolve(Object.entries({ top_level_dead_promise_resolve_object_entries_value_key: "top_level_dead_promise_resolve_object_entries_value" })[0][1]);
 Promise.resolve(Object.entries(["top_level_dead_promise_resolve_object_entries_array_value"])[0][1]);
@@ -921,6 +927,9 @@ function usedLocal(value: number): number {
     Promise.resolve(Object.values(["local_dead_promise_resolve_object_values_array_join"]).join("|"));
     Promise.resolve(Object.values("local_dead_promise_resolve_object_values_string_to_string").toString());
     Promise.resolve(Object.values({ local_dead_promise_resolve_object_values_object_join: "local_dead_promise_resolve_object_values_object_join" }).join("|"));
+    Promise.resolve(Object.values(Object.assign({} as { local_dead_promise_resolve_object_values_assign_join: string }, { local_dead_promise_resolve_object_values_assign_join: "local_dead_promise_resolve_object_values_assign_join" })).join("|"));
+    Promise.resolve(Object.values(Object.fromEntries<{ local_dead_promise_resolve_object_values_from_entries_join: string }>([["local_dead_promise_resolve_object_values_from_entries_join", "local_dead_promise_resolve_object_values_from_entries_join"]])).join("|"));
+    Promise.resolve(Object.values(Object.fromEntries<{ local_dead_promise_resolve_object_values_entries_from_entries_join: string }>(Object.entries({ local_dead_promise_resolve_object_values_entries_from_entries_join: "local_dead_promise_resolve_object_values_entries_from_entries_join" }))).join("|"));
     Promise.resolve(Object.entries({ local_dead_promise_resolve_object_entries_key: "local_dead_promise_resolve_object_entries_value" })[0][0]);
     Promise.resolve(Object.entries({ local_dead_promise_resolve_object_entries_value_key: "local_dead_promise_resolve_object_entries_value" })[0][1]);
     Promise.resolve(Object.entries(["local_dead_promise_resolve_object_entries_array_value"])[0][1]);